requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
click==8.1.7
dj-database-url==2.1.0
djangorestframework==3.14.0
Django==4.2.11
fastapi==0.116.1
gunicorn==21.2.0
h11==0.16.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.9
python-dotenv=1.0.0
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
whitenoise==6.6.0
